// src/components/types.ts
export type ExperienceItem = {
    period: string;
    role: string;
    company: string;
    companyUrl: string;
    points: string[];
    technologies: string[];
  };
  
  export type ProjectItem = {
    title: string;
    description: string;
    imageUrl: string;
    projectUrl: string;
    technologies: string[];
  };
  
  export type WritingItem = {
    year: string;
    title: string;
    description: string;
    imageUrl: string;
    articleUrl: string;
  };